02_intent_router/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain_openai>=1.1.13
pydantic>=2.7.0
pytest>=8.0.0
typer>=0.12.0
rich>=13.7.0
python-dotenv>=1.0.0